move all knobs to KnobFactory and update index.ts
