+++ packages/audio-components/src/elements/RecordButton.ts
@@ -3,7 +3,10 @@ import van, { State } from '@repo/vanjs-core';
 import { ElementProps } from '@repo/vanjs-core/element';
 import { createAudioRecorder, type Recorder } from '@repo/audiolib';
 import { getSampler, onRegistryChange } from '../SamplerRegistry';
-import { BUTTON_STYLE, BUTTON_ACTIVE_STYLE } from './ComponentStyles';
+import {
+  BUTTON_STYLE,
+  BUTTON_ACTIVE_STYLE,
+} from '../shared/styles/component-styles';
 
 const { div, button } = van.tags;
