+++ packages/audio-components/src/elements/ToggleComponents.ts
@@ -3,7 +3,7 @@ import van, { State } from '@repo/vanjs-core';
 import { ElementProps } from '@repo/vanjs-core/element';
 import { getSampler } from '../SamplerRegistry';
 import { Toggle } from './primitives/VanToggle';
-import { INLINE_COMPONENT_STYLE } from './ComponentStyles';
+import { INLINE_COMPONENT_STYLE } from '../shared/styles/component-styles';
 
 const { div, label } = van.tags;
