+++ packages/audio-components/src/shared/utils/visual/generateColors.ts
