+++ packages/audio-components/src/shared/utils/visual/waveform-utils.ts
